[filterpacket/filterintrin/vm_opcodes] Rework filter VM and introduce complex AS PATH filter logic
Now the filter VM is able to leave packet iterators open between
opcodes, allowing complex AS PATH matching logic (e.g. a path starts
with an AS and ends with another AS).
This also enables the VM to perform other optimizations.
The VM now has a different way to handle the BLK instruction, which
simplifies the structure and behaves well with nested BLKs.
The general code in filterintrin.c has been cleaned up a bit.